fix: search results on accented card names
Uses third-party module 'regex' for its support of Unicode categories.

The bug caused fuzzy string search in /gacha view to not return a match on accented names, such as "Déjà Vu", as the search does not recognize "e" to be the same as "é".

This fix removes accents and the like from card names for searching. Similar processing is enforced on the search key.