fix: remove extra btns in /server emoji due to timeout bug
Mitsuki changed the message the paginator was in. However, this original message is still referenced by the interactions.py paginator for their timeout. As such, subsequent messages may timeout prematurely.

A fix would necessitate rewriting a good chunk of Paginator code to accommodate a global timeout store and would be done in a later version.